SRT passing test case 1
Issues with test case 2
test case 3 not checked

+++ tau_calc.h
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <math.h>
 
-int calc_tau(double prev_tau, int actual, double alpha)
+int calc_tau(int prev_tau, int actual, double alpha)
 {
     /* recalculate tau based on previous tau, alpha, and actual time */
     // double new_tau = alpha * actual + (1.0 - alpha) * prev_tau;
